requirements-dev.txt
mock
moto
nose
django_nose
nose-testconfig

